[Translation] ResourceBundleLoader to IcuRes/DatFileLoader
