+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/translation.xml
@@ -15,7 +15,8 @@
         <parameter key="translation.loader.mo.class">Symfony\Component\Translation\Loader\MoFileLoader</parameter>
         <parameter key="translation.loader.qt.class">Symfony\Component\Translation\Loader\QtTranslationsLoader</parameter>
         <parameter key="translation.loader.csv.class">Symfony\Component\Translation\Loader\CsvFileLoader</parameter>
-        <parameter key="translation.loader.rb.class">Symfony\Component\Translation\Loader\ResourceBundleLoader</parameter>
+        <parameter key="translation.loader.res.class">Symfony\Component\Translation\Loader\IcuResFileLoader</parameter>
+        <parameter key="translation.loader.dar.class">Symfony\Component\Translation\Loader\IcuDatFileLoader</parameter>
         <parameter key="translation.loader.ini.class">Symfony\Component\Translation\Loader\IniFileLoader</parameter>
         <parameter key="translation.dumper.php.class">Symfony\Component\Translation\Dumper\PhpFileDumper</parameter>
         <parameter key="translation.dumper.xliff.class">Symfony\Component\Translation\Dumper\XliffFileDumper</parameter>
@@ -25,6 +26,7 @@
         <parameter key="translation.dumper.qt.class">Symfony\Component\Translation\Dumper\QtFileDumper</parameter>
         <parameter key="translation.dumper.csv.class">Symfony\Component\Translation\Dumper\CsvFileDumper</parameter>
         <parameter key="translation.dumper.ini.class">Symfony\Component\Translation\Dumper\IniFileDumper</parameter>
+        <parameter key="translation.dumper.res.class">Symfony\Component\Translation\Dumper\IcuResFileDumper</parameter>
         <parameter key="translation.extractor.php.class">Symfony\Bundle\FrameworkBundle\Translation\PhpExtractor</parameter>
         <parameter key="translation.loader.class">Symfony\Bundle\FrameworkBundle\Translation\TranslationLoader</parameter>
         <parameter key="translation.extractor.class">Symfony\Component\Translation\Extractor\ChainExtractor</parameter>
@@ -76,8 +78,12 @@
             <tag name="translation.loader" alias="csv" />
         </service>
 
-        <service id="translation.loader.rb" class="%translation.loader.rb.class%">
-            <tag name="translation.loader" alias="rb" />
+        <service id="translation.loader.res" class="%translation.loader.res.class%">
+            <tag name="translation.loader" alias="res" />
+        </service>
+
+        <service id="translation.loader.dat" class="%translation.loader.res.class%">
+            <tag name="translation.loader" alias="dat" />
         </service>
 
         <service id="translation.loader.ini" class="%translation.loader.ini.class%">
@@ -114,6 +120,10 @@
 
         <service id="translation.dumper.ini" class="%translation.dumper.ini.class%">
             <tag name="translation.dumper" alias="ini" />
+        </service>
+
+        <service id="translation.dumper.res" class="%translation.dumper.res.class%">
+            <tag name="translation.dumper" alias="res" />
         </service>
 
         <service id="translation.extractor.php" class="%translation.extractor.php.class%">
