+++ src/Symfony/Component/Translation/Loader/IcuDatFileLoader.php
@@ -1,0 +1,44 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Translation\Loader;
+
+use Symfony\Component\Translation\MessageCatalogue;
+use Symfony\Component\Config\Resource\FileResource;
+
+/**
+ * IcuResFileLoader loads translations from a resource bundle.
+ *
+ * @author stealth35
+ */
+class IcuDatFileLoader extends IcuResFileLoader
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function load($resource, $locale, $domain = 'messages')
+    {
+        $rb = new \ResourceBundle($locale, $resource);
+
+        if (!$rb) {
+            throw new \RuntimeException("cannot load this resource : $resource");
+        } elseif (intl_is_failure($rb->getErrorCode())) {
+            throw new \RuntimeException($rb->getErrorMessage(), $rb->getErrorCode());
+        }
+
+        $messages = $this->flatten($rb);
+        $catalogue = new MessageCatalogue($locale);
+        $catalogue->add($messages, $domain);
+        $catalogue->addResource(new FileResource($resource.'.dat'));
+
+        return $catalogue;
+    }
+}
